Add MoveCommand
The only thing I don't like about this refactoring so far is the
duplication of the ``` in the strings, which previously was handled in
the bot code which felt nice.

I'd like to get it in there somehow, but it's hard to figure out how
you'd do that without not returning a string. I suppose one (maybe
weird) way of doing it would be to always send the status back. Another
way of doing it could be to put in some hooks or marker on the commands
as to whether they should send back a board status or not. Which could
then be wrapped.