Refactor TicTacBot to use bot commands
This was done in multiple small steps, changing the if (command.equals
("")) type strings into botCommand.canHandle with each sub type being
used. Once that was complete, I noticed we needed to do the game over
check in the middle of things, before the move or show command, so I
made two command lists and added the loop.

Plenty of the helper methods can now be removed since their
functionality has moved into the various commands.

As noted in a prior commit, I want to figure out a better way to wrap
things with the ``` and not repeat that various places. I might just add
 it to the toString for the game and call it a day though.